sessions with token auth fixed

+++ authentication/function.php
@@ -5,45 +5,46 @@ class func {
 
     public static function checkLoginState($dbh){
 
-        if(!isset($_SESSION['id']) || !isset($_COOKIE['PHPSESSID'])){
+        if(!isset($_SESSION['user_id']) || !isset($_COOKIE['PHPSESSID'])){
             session_start();
         }
-        if(isset($_COOKIE['id']) && isset($_COOKIE['token']) && isset($_COOKIE['serial'])){
+        if(isset($_COOKIE['user_id']) && isset($_COOKIE['token']) && isset($_COOKIE['serial'])){
             
             $query = "SELECT * From sessions Where user_id = :userid AND token = :token AND serial = :serial;";
 
-            $userid = $_COOKIE['userid'];
+            $userid = $_COOKIE['user_id'];
             $token = $_COOKIE['token'];
             $serial = $_COOKIE['serial'];
 
             $stmt = $dbh->prepare($query);
             $stmt->execute(array(':userid' => $userid, ':token' => $token, 'serial' => $serial));
-
+ 
             $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
-            if($row['userid'] > 0){
+           if($row['id'] > 0){
 
                 if(
-                    $row['userid'] == $_COOKIE['userid'] &&
-                    $row['userid'] == $_COOKIE['token'] &&
-                    $row['userid'] == $_COOKIE['serial'] 
+                    $row['user_id'] == $_COOKIE['user_id'] &&
+                    $row['token'] == $_COOKIE['token'] &&
+                    $row['serial'] == $_COOKIE['serial'] 
                 ){
                     if(
-                        $row['userid'] == $_SESSION['userid'] &&
-                        $row['userid'] == $_SESSION['token'] &&
-                        $row['userid'] == $_SESSION['serial'] 
+                        $row['user_id'] == $_SESSION['user_id'] &&
+                        $row['token'] == $_SESSION['token'] &&
+                        $row['serial'] == $_SESSION['serial'] 
                     ){
                         return true;
                     }
                 }
 
-            }
+            } 
 
         }
     }
     public static function createRecord($dbh, $id, $username){
 
-        $dbh->prepare('DELETE From sessions WHERE userid');
+        $query = 'INSERT Into sessions (user_id, token, serial, date ) Values (:user_id, :token, :serial, "22/04/2020")';
+        $dbh->prepare("DELETE From sessions WHERE user_id = :userid;")->execute(array(':userid' => $id));
 
 
         $token = func::createString(32);
@@ -51,20 +52,23 @@ class func {
         func::createCookie($username, $id, $token, $serial);
         func::createSession($username, $id, $token, $serial);
 
-        //$dbh->prepare('INSERT Into sessions () Values WHERE userid'); Part 3 Min 11
-
+        $stmt = $dbh->prepare($query);
+        $stmt->execute(array(':user_id' => $id, ':token' => $token, ':serial' => $serial));
     }
     public static function createCookie($username, $id, $token, $serial){
-        setcookie('id', $id, time() + (86400) * 30, "/");
+        setcookie('user_id', $id, time() + (86400) * 30, "/");
         setcookie('username', $username, time() + (86400) * 30, "/");
         setcookie('token', $token, time() + (86400) * 30, "/");
         setcookie('serial', $serial, time() + (86400) * 30, "/");
     }
 
     public static function createSession($username, $id, $token, $serial){
-        if(isset($_SESSION['id']) || !isset($_COOKIE['PHPSESSID'])){
+        if(isset($_SESSION['user_id']) || !isset($_COOKIE['PHPSESSID'])){
             session_start();
         }
+        $_SESSION['user_id'] = $id;
+        $_SESSION['token'] = $token;
+        $_SESSION['serial'] = $serial;
         $_SESSION['username'] = $username;
         
     }
